Conflicts:
	Data/level_1/cameras.xml
	MasterUAB/Core/Engine/UABEngine.cpp
	MasterUAB/Core/Engine/UABEngine.h
	MasterUAB/Graphics/Camera/Camera.h
	MasterUAB/Graphics/Camera/CameraController.h
	MasterUAB/Graphics/Camera/CameraControllerManager.cpp
	MasterUAB/Graphics/Camera/CameraControllerManager.h
	MasterUAB/Graphics/Camera/CameraInfo.cpp
	MasterUAB/Graphics/Camera/CameraKeyController.cpp
	MasterUAB/Graphics/Graphics.vcxproj.filters
	MasterUAB/VideogameTest/Application.cpp

JOAN NO LA LIES!

+++ MasterUAB/Core/InputManager/InputManagerImplementation.cpp
@@ -174,7 +174,7 @@ void CInputManagerImplementation::LoadCommandsFromFile(const std::string& path)
 					{
 					case GAMEPAD:
 						axis.gamepad.axis = ParseGamepadAxis(l_Axis.GetPszProperty("gamepad_axis", "X"));
-						axis.gamepad.gamepadNumber = l_Axis.GetFloatProperty("gamepad_number", 0, false);
+						axis.gamepad.gamepadNumber = l_Axis.GetIntProperty("gamepad_number", 0, false);
 						break;
 					case MOUSE:
 						axis.mouse.axis = ParseMouseAxis(l_Axis.GetPszProperty("mouse_axis", "X"));
